+++ JTSim/Simulator/Solvers/SolverRK4Enhanced.cs
@@ -9,14 +9,14 @@ namespace JTSim
 {
     public class SolverRK4Enhanced : ISolver
     {
-        public Vector<float> Solve(ContinousModel model, Vector<float> state, float input, float t, float h)
+        public Vector<double> Solve(ContinousModel model, Vector<double> state, double input, double t, double h)
         {
-            float third = 1f / 3f;
-            Vector<float> k1 = h * model.DifferentialEquasions(state, input, t);
-            Vector<float> k2 = h * model.DifferentialEquasions(state + third * k1, input, t + third * h);
-            Vector<float> k3 = h * model.DifferentialEquasions(state + (-third * k1 + k2), input, t + 2f * third * h);
-            Vector<float> k4 = h * model.DifferentialEquasions(state + (k1 - k2 + k3), input, t + h);
-            return state + 1f / 8f * (k1 + 3f * k2 + 3f * k3 + k4);
+            double third = 1d / 3d;
+            Vector<double> k1 = h * model.DifferentialEquasions(state, input, t);
+            Vector<double> k2 = h * model.DifferentialEquasions(state + third * k1, input, t + third * h);
+            Vector<double> k3 = h * model.DifferentialEquasions(state + (-third * k1 + k2), input, t + 2d * third * h);
+            Vector<double> k4 = h * model.DifferentialEquasions(state + (k1 - k2 + k3), input, t + h);
+            return state + 1d / 8d * (k1 + 3f * k2 + 3f * k3 + k4);
         }
     }
 }
