+++ JTSim/Simulator/Systems/Discrete/DiscreteSystem.cs
@@ -5,20 +5,20 @@ namespace JTSim
     public class DiscreteSystem : GenericSystem
     {
         public DiscreteModel model;
-        public Vector<float> states;
-        public Vector<float> inputs;
+        public Vector<double> states;
+        public Vector<double> inputs;
         private TransportDelay transportDelay;
-        private float delay;
+        private double delay;
 
-        public DiscreteSystem(DiscreteModel model) : this(model, 0f) { }
+        public DiscreteSystem(DiscreteModel model) : this(model, 0d) { }
 
-        public DiscreteSystem (DiscreteModel model, float delay)
+        public DiscreteSystem (DiscreteModel model, double delay)
         {
             this.model = model;
             this.delay = delay;
         }
 
-        public override void Step(float u, float t, float h)
+        public override void Step(double u, double t, double h)
         {
             PushToVectorF(inputs, u);
             PushToVectorF(states, model.DifferenceEquasion(states, inputs, t, h));
@@ -27,16 +27,16 @@ namespace JTSim
             output = transportDelay.Step(output);
         }
 
-        public override void Init(float h)
+        public override void Init(double h)
         {
             states = model.initStates.Clone();
             inputs = model.initInputs.Clone();
-            float initOutput = model.OutputEquation(states, inputs);
+            double initOutput = model.OutputEquation(states, inputs);
             transportDelay = new TransportDelay(delay, initOutput, h);
             output = initOutput;
         }
 
-        private void PushToVectorF (Vector<float> vect, float newVal)
+        private void PushToVectorF (Vector<double> vect, double newVal)
         {
             for (int i = vect.Count - 2; i >= 0; i--)
                 vect[i + 1] = vect[i];
